+++ frontend/src/App.jsx
@@ -6,12 +6,14 @@ import Navdoctor from './components/Navbar/Navdoctor';
 import { Routes, Route, useLocation } from 'react-router-dom';
 import Login from './components/Login-page/Login';
 import Register from './components/Register-page/Register';
+// import ViewRecord from './components/ViewRecord/ViewRecord';
 import Home from './components/Home/Home';
 import Footer from './components/Footer/Footer';
 import Dashboard from './components/User-Upload-Dashboard/userDashboard';
 
 function App() {
   const location = useLocation();
+  const isHome = location.pathname === '/';
   const [showFooter, setShowFooter] = useState(false);
 
   useEffect(() => {
@@ -37,7 +39,7 @@ function App() {
 
   return (
     <div className="App">
-      {<Navdoctor/>}
+      {!isHome && <Navdoctor />}
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/login" element={<Login />} />
@@ -45,9 +47,13 @@ function App() {
         <Route path='/' element={<Home />} />
         <Route path="/navpatient" element={<Navpatient />} />
         <Route path="/navdoctor" element={<Navdoctor />} />
+<<<<<<< HEAD
         <Route path="/upload" element={<Dashboard />} />
+=======
+        {/* <Route path="/viewrecord" element={<ViewRecord />} /> */}
+>>>>>>> fbb73af5ac8b1851069a982f39d1c756cad14025
       </Routes>
-      { <Footer />}
+      {!isHome && <Footer />}
     </div>
   );
 }